refactor(lifecycle): consolidate lifecycle management and bootstrap services
