--- backend/app/init_data/__init__.py
@@ -1,3 +1,0 @@
-from app.init_data.bootstrap import DataBootstrapService
-
-__all__ = ["DataBootstrapService"]
