--- backend/app/init_data/mcp_servers/__init__.py
@@ -1,3 +1,0 @@
-from app.init_data.mcp_servers.builtin_mcp import McpServerBootstrapService
-
-__all__ = ["McpServerBootstrapService"]
