--- backend/app/init_data/skills/__init__.py
@@ -1,11 +1,0 @@
-from app.init_data.skills.builtin_skills import (
-    BUILTIN_SKILLS,
-    SYSTEM_SKILL_OWNER_USER_ID,
-    SkillBootstrapService,
-)
-
-__all__ = [
-    "BUILTIN_SKILLS",
-    "SYSTEM_SKILL_OWNER_USER_ID",
-    "SkillBootstrapService",
-]
