+++ backend/app/lifecycle/__init__.py
@@ -1,0 +1,17 @@
+from app.lifecycle.bootstrap import LifecycleBootstrapService
+from app.lifecycle.builtin_mcp import McpServerBootstrapService
+from app.lifecycle.builtin_skills import (
+    BUILTIN_SKILLS,
+    SYSTEM_SKILL_OWNER_USER_ID,
+    SkillBootstrapService,
+)
+from app.lifecycle.lifespan import lifespan
+
+__all__ = [
+    "BUILTIN_SKILLS",
+    "SYSTEM_SKILL_OWNER_USER_ID",
+    "LifecycleBootstrapService",
+    "McpServerBootstrapService",
+    "SkillBootstrapService",
+    "lifespan",
+]
