+++ backend/app/lifecycle/bootstrap.py
@@ -1,18 +1,18 @@
 import logging
 
 from app.core.database import SessionLocal
-from app.init_data.mcp_servers.builtin_mcp import McpServerBootstrapService
-from app.init_data.skills.builtin_skills import SkillBootstrapService
+from app.lifecycle.builtin_mcp import McpServerBootstrapService
+from app.lifecycle.builtin_skills import SkillBootstrapService
 
 logger = logging.getLogger(__name__)
 
 
-class DataBootstrapService:
-    """Bootstrap built-in data that must exist before serving requests."""
+class LifecycleBootstrapService:
+    """Bootstrap built-in data required during application startup."""
 
     @classmethod
     def bootstrap_all(cls) -> None:
-        """Bootstrap all built-in data with a dedicated database session."""
+        """Bootstrap all built-in resources with a dedicated database session."""
         db = SessionLocal()
         try:
             SkillBootstrapService.bootstrap_builtin_skills(db)
@@ -20,7 +20,7 @@ class DataBootstrapService:
             db.commit()
         except Exception:
             db.rollback()
-            logger.exception("bootstrap_all_failed")
+            logger.exception("lifecycle_bootstrap_failed")
             raise
         finally:
             db.close()
