+++ backend/app/lifecycle/builtin_mcp.py
